requirements_streamlit.txt
streamlit
google-generativeai
python-dotenv
python-docx
pandas